[refactor]: game component들 패키징 변경
